requirements.txt
pytest==6.0.1
coverage==5.2.1
pylint==2.5.3
Flask==1.1.2
hypothesis==5.24.3
requests==2.24.0
virtualenv==20.0.31
pylint-venv==2.1.1
mypy==0.782
PyJWT==1.7.1
attrs==20.2.0
email-validator==1.1.1
flask-cors==3.0.9
blinker==1.4
werkzeug==1.0.1
wikiquote==0.1.14
Pillow==8.0.1
psycopg2==2.8.6
gunicorn==20.0.4
